fix(auth): preserve client portal branding through password reset
