+++ packages/auth/src/components/ClientPortalSignIn.tsx
@@ -20,9 +20,10 @@ type TenantBranding = {
 
 interface ClientPortalSignInProps {
   branding?: TenantBranding | null;
+  portalDomain?: string;
 }
 
-export default function ClientPortalSignIn({ branding }: ClientPortalSignInProps) {
+export default function ClientPortalSignIn({ branding, portalDomain }: ClientPortalSignInProps) {
   const { t } = useTranslation('client-portal');
   const [isAlertOpen, setIsAlertOpen] = useState(false);
   const [alertInfo, setAlertInfo] = useState<AlertProps>({ type: 'success', title: '', message: '' });
@@ -285,6 +286,7 @@ export default function ClientPortalSignIn({ branding }: ClientPortalSignInProps
                 onError={handleError}
                 onTwoFactorRequired={() => setIsOpen2FA(true)}
                 tenantSlug={tenantSlug}
+                portalDomain={portalDomain}
               />
               <div className="mt-6 pt-6 border-t text-center">
                 <a href="/auth/msp/signin" className="text-sm text-[rgb(var(--color-text-600))] hover:text-[rgb(var(--color-primary-500))]">
